[TT-13670] Decouple OAuthManager behind interface (#6735)
### **User description**
https://tyktech.atlassian.net/browse/TT-13670


___

### **PR Type**
Enhancement, Tests


___

### **Description**
- Refactored `OAuthManager` to introduce a new `Storage()` method,
encapsulating access to `OsinServer.Storage`.
- Introduced `OAuthManagerInterface` to decouple `OAuthManager` from its
concrete implementation.
- Updated `APISpec` to use the `OAuthManagerInterface` instead of the
concrete `OAuthManager`.
- Refactored multiple files to replace direct access to
`OsinServer.Storage` with the new `Storage()` method.
- Relocated `URLSpec` struct from `api_definition.go` to
`model_urlspec.go` for better modularity.
- Updated related test cases to align with the new `Storage()` method
and interface changes.



___



### **Changes walkthrough** 📝
<table><thead><tr><th></th><th align="left">Relevant
files</th></tr></thead><tbody><tr><td><strong>Enhancement</strong></td><td><table>
<tr>
  <td>
    <details>
<summary><strong>api.go</strong><dd><code>Refactor OAuthManager to use
Storage() method in API functions</code></dd></summary>
<hr>

gateway/api.go

<li>Replaced direct access to <code>OsinServer.Storage</code> with a new
<code>Storage()</code> <br>method in <code>OAuthManager</code>.<br> <li>
Updated multiple functions to use the new <code>Storage()</code> method
for OAuth <br>client operations.<br>


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6735/files#diff-644cda3aeb4ac7f325359e85fcddb810f100dd5e6fa480b0d9f9363a743c4e05">+16/-17</a>&nbsp;
</td>

</tr>

<tr>
  <td>
    <details>
<summary><strong>api_definition.go</strong><dd><code>Introduce
OAuthManagerInterface and update APISpec</code>&nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

gateway/api_definition.go

<li>Introduced <code>OAuthManagerInterface</code> interface with a
<code>Storage()</code> method.<br> <li> Updated <code>APISpec</code> to
use <code>OAuthManagerInterface</code> instead of a concrete
<br><code>OAuthManager</code>.<br>


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6735/files#diff-0cf80174bbafb36f6d4f4308ebbd971b2833b76a936bad568220aa1a4ba0ee8b">+8/-37</a>&nbsp;
&nbsp; </td>

</tr>

<tr>
  <td>
    <details>
<summary><strong>model_urlspec.go</strong><dd><code>Relocate URLSpec
struct to model_urlspec.go</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
</dd></summary>
<hr>

gateway/model_urlspec.go

<li>Moved <code>URLSpec</code> struct definition from
<code>api_definition.go</code> to <br><code>model_urlspec.go</code>.<br>
<li> No functional changes, purely structural refactoring.<br>


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6735/files#diff-ed766dd1814a557a1943cd3483c6ef511ad1b8febc7bd59982792d0ab7e23cff">+41/-0</a>&nbsp;
&nbsp; </td>

</tr>

<tr>
  <td>
    <details>
<summary><strong>mw_jwt.go</strong><dd><code>Update JWT middleware to
use OAuthManager Storage() method</code></dd></summary>
<hr>

gateway/mw_jwt.go

<li>Updated JWT middleware to use the new <code>Storage()</code> method
for OAuth <br>client retrieval.<br>


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6735/files#diff-e8bce0f6790c8c56b30e24dbeebb0fc4aa0879ab5ea5f6ef6dbe68931410e237">+1/-1</a>&nbsp;
&nbsp; &nbsp; </td>

</tr>

<tr>
  <td>
    <details>
<summary><strong>mw_oauth2_key_exists.go</strong><dd><code>Update OAuth2
middleware to use OAuthManager Storage() method</code></dd></summary>
<hr>

gateway/mw_oauth2_key_exists.go

<li>Updated OAuth2 key existence middleware to use the new
<code>Storage()</code> <br>method for OAuth client retrieval.<br>


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6735/files#diff-c50b0affcb835d15ff9bd169b21775b8023ebb0558a0c59d6c26cb821db544e9">+1/-1</a>&nbsp;
&nbsp; &nbsp; </td>

</tr>

<tr>
  <td>
    <details>
<summary><strong>oauth_manager.go</strong><dd><code>Add Storage() method
to OAuthManager</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

gateway/oauth_manager.go

<li>Added <code>Storage()</code> method to <code>OAuthManager</code> to
encapsulate access to <br><code>OsinServer.Storage</code>.<br>


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6735/files#diff-bcb7d4bc0ecfd071eff1cdbc75ccad0f5debef1d97f372c41c3637174b7448f9">+4/-0</a>&nbsp;
&nbsp; &nbsp; </td>

</tr>
</table></td></tr><tr><td><strong>Tests</strong></td><td><table>
<tr>
  <td>
    <details>
<summary><strong>oauth_manager_test.go</strong><dd><code>Update
OAuthManager tests to use Storage() method</code>&nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

gateway/oauth_manager_test.go

<li>Updated tests to use the new <code>Storage()</code> method for
setting OAuth <br>clients.<br>


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6735/files#diff-139c3beb238f234728bde9f619f501cf730a7e275d17c8511277272d1dcd6fc6">+4/-4</a>&nbsp;
&nbsp; &nbsp; </td>

</tr>

<tr>
  <td>
    <details>
<summary><strong>rpc_storage_handler_test.go</strong><dd><code>Update
RPC storage handler tests to use Storage() method</code>&nbsp;
</dd></summary>
<hr>

gateway/rpc_storage_handler_test.go

<li>Updated test cases to use the new <code>Storage()</code> method for
OAuth client <br>operations.<br>


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6735/files#diff-69de989a02b3bc32ae376c514ee84633c609200db22385c0e16c361d6ea74cd6">+1/-1</a>&nbsp;
&nbsp; &nbsp; </td>

</tr>
</table></td></tr></tr></tbody></table>

___

> 💡 **PR-Agent usage**: Comment `/help "your question"` on any pull
request to receive relevant information

---------
