+++ gateway/oauth_manager.go
@@ -342,6 +342,10 @@ type OAuthManager struct {
 	Gw         *Gateway `json:"-"`
 }
 
+func (o *OAuthManager) Storage() ExtendedOsinStorageInterface {
+	return o.OsinServer.Storage
+}
+
 // HandleAuthorisation creates the authorisation data for the request
 func (o *OAuthManager) HandleAuthorisation(r *http.Request, complete bool, session string) *osin.Response {
 	resp := o.OsinServer.NewResponse()
